Moved relatedproducts out
